video-clip-contextualizer/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.35.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
sentence-transformers>=2.2.0
redis>=5.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.0.0
moviepy>=1.0.3
librosa>=0.10.0
soundfile>=0.12.0
psutil>=5.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pyyaml>=6.0
scipy>=1.11.0
spacy>=3.7.0
datasets>=2.14.0